requirements_jetson.txt
# ============================================================
# Requirements for KV-Cache experiments on Jetson Orin NX
# ============================================================
#
# PyTorch: install the Jetson-specific wheel from NVIDIA:
#   https://forums.developer.nvidia.com/t/pytorch-for-jetson/
# Example (JetPack 6.x / CUDA 12.x):
#   pip install torch-2.x.x+nv... (from NVIDIA wheel)
#
# vLLM: see scripts/setup_vllm_jetson.sh
# KIVI: see scripts/setup_kivi_jetson.sh
# ============================================================

# --- Core inference ---
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
safetensors>=0.4.0
sentencepiece>=0.2.0

# --- Config & data ---
pyyaml>=6.0
pandas>=2.0.0

# --- Visualization ---
matplotlib>=3.7.0
seaborn>=0.13.0

# --- Utilities ---
tqdm>=4.66.0
numpy>=1.24.0
